unify needless files
